+++ lib/src/compiler/evaluation/operators/plus.cc
@@ -31,20 +31,20 @@ namespace puppet { namespace compiler { namespace evaluation { namespace operato
             return left + right;
         }
 
-        result_type operator()(int64_t left, long double right) const
+        result_type operator()(int64_t left, double right) const
         {
-            return operator()(static_cast<long double>(left), right);
+            return operator()(static_cast<double>(left), right);
         }
 
-        result_type operator()(long double left, int64_t right) const
+        result_type operator()(double left, int64_t right) const
         {
-            return operator()(left, static_cast<long double>(right));
+            return operator()(left, static_cast<double>(right));
         }
 
-        result_type operator()(long double left, long double right) const
+        result_type operator()(double left, double right) const
         {
             feclearexcept(FE_OVERFLOW | FE_UNDERFLOW);
-            long double result = left + right;
+            double result = left + right;
             if (fetestexcept(FE_OVERFLOW)) {
                 throw evaluation_exception((boost::format("addition of %1% and %2% results in an arithmetic overflow.") % left % right).str(), _context.right_context());
             } else if (fetestexcept(FE_UNDERFLOW)) {
@@ -140,7 +140,7 @@ namespace puppet { namespace compiler { namespace evaluation { namespace operato
         template <
             typename Right,
             typename = typename enable_if<!is_same<Right, int64_t>::value>::type,
-            typename = typename enable_if<!is_same<Right, long double>::value>::type
+            typename = typename enable_if<!is_same<Right, double>::value>::type
         >
         result_type operator()(int64_t, Right const& right) const
         {
@@ -150,9 +150,9 @@ namespace puppet { namespace compiler { namespace evaluation { namespace operato
         template <
             typename Right,
             typename = typename enable_if<!is_same<Right, int64_t>::value>::type,
-            typename = typename enable_if<!is_same<Right, long double>::value>::type
+            typename = typename enable_if<!is_same<Right, double>::value>::type
         >
-        result_type operator()(long double, Right const& right) const
+        result_type operator()(double, Right const& right) const
         {
             throw evaluation_exception((boost::format("expected %1% for arithmetic addition but found %2%.") % types::numeric::name() % value(right).get_type()).str(), _context.right_context());
         }
